AI_powered_personal_stylist/backend/requirements.txt
flask
google-generativeai
python-dotenv
requests
Pillow